Moved AWS CLI function to iam.rs and include it in menu Menu changes reflect AWS CLI
Considering deleting util function, moving away from that text menu
style

Changed install script with setting Rust enviornment
