+++ src/iam.rs
@@ -5,6 +5,8 @@ use aws_sdk_iam::operation::{
 };
 // use aws_sdk_iam::model::{ListUserPoliciesInput, ListUserPoliciesOutput};
 
+use std::{env, process::Command};
+use std::io::{self, Write};
 use aws_sdk_iam::error::SdkError;
 use tokio::time::{sleep, Duration};
 use aws_sdk_iam::types::{AccessKey, Policy, PolicyScopeType, Role, User};
@@ -54,7 +56,6 @@ pub async fn list_groups_for_user(
 pub async fn is_admin_user(user: &User, iam_client: &Client) -> bool {
     let user_name = user.user_name.clone(); 
 
-    // Check if the user belongs to the admin group
     let result = list_groups_for_user(iam_client, user_name.clone(), None).await;
     println!("Result: {:?}", result);
     match result {
@@ -84,3 +85,51 @@ pub async fn get_user(
     let response: GetUserOutput = client.get_user().send().await?;
     Ok(response)
 }
+
+
+// Asynchronously runs the Administrator AWS CLI if the user is an admin
+pub async fn run_admin_cli() -> Result<(), Error> { 
+
+    let iam_client = start_iam().await;
+    match iam_client {
+        Ok(client) => {
+            let user = get_user(&client).await.unwrap();
+            if is_admin_user(&user.user.as_ref().unwrap(), &client).await {
+                let stdin = io::stdin();
+                let mut stdout = io::stdout();
+        
+                println!("\nRunning Administrator AWS CLI:");
+                loop {
+                    print!("aws> ");
+                    stdout.flush().unwrap();
+                    let mut input = String::new();
+                    stdin.read_line(&mut input).unwrap();
+                    if input.trim().eq_ignore_ascii_case("exit") {
+                        break;
+                    }
+                    let args: Vec<&str> = input.trim().split_whitespace().collect();
+                    if args.is_empty() {
+                        continue;
+                    }
+        
+                    let output = Command::new("aws")
+                        .args(&args)
+                        .output()
+                        .expect("Failed to execute command");
+        
+                    io::stdout().write_all(&output.stdout).unwrap();
+                    io::stderr().write_all(&output.stderr).unwrap();
+                } 
+            }
+            else{
+                println!("\n{} is not an admininstrator!", user.user.as_ref().unwrap().user_name);
+                println!("Please contact the admininistrator for more information.");
+            }
+        }
+        Err(err) => {
+            println!("Error occurred starting IAM client: {}", err);
+        }
+    }
+    Ok(())
+}
+
